// src/redux/types.ts
export enum WEB3_ACTIONS {
    SET_WALLET_IS_CONNECTED = 'SET_IS_CONNECTED',
    SET_STAKING_TOKEN_DATA = 'SET_STAKING_TOKEN_DATA',
    SET_REWARD_TOKEN_DATA = 'SET_REWARD_TOKEN_DATA',
    SET_STAKING_CONTRACT_DATA = 'SET_STAKING_CONTRACT_DATA',
    SET_TRANSACTIONS = 'SET_TRANSACTIONS',
}

export enum WEB3_GETTERS {
    GET_WALLET_IS_CONNECTED = 'GET_IS_CONNECTED',
    GET_STAKING_TOKEN = 'GET_STAKING_TOKEN',
    GET_REWARD_TOKEN = 'GET_REWARD_TOKEN',
    GET_STAKING_CONTRACT_DATA = 'GET_STAKING_CONTRACT_DATA',
    GET_TRANSACTIONS = 'GET_TRANSACTIONS',
}